% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{getCoef}
\alias{getCoef}
\alias{getCoef,LinRegP-method}
\title{Set the generic for accessor (getter) for coefficients}
\usage{
getCoef(object)

\S4method{getCoef}{LinRegP}(object)
}
\arguments{
\item{object}{object of class LinRegP}
}
\description{
Set the generic for accessor (getter) for coefficients
}
\section{Methods (by class)}{
\itemize{
\item \code{LinRegP}: 
}}

