% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\docType{methods}
\name{initialize,LinRegP-method}
\alias{initialize,LinRegP-method}
\title{Initialiser for LinRegP objects}
\usage{
\S4method{initialize}{LinRegP}(.Object, model, intercept, outcome, covariates,
  modelCoef, fittedVals, residuals, summaryStats, leverage)
}
\arguments{
\item{.Object}{object of class LinRegP}

\item{model}{character string representing model}

\item{intercept}{logical. is intercept included?}

\item{outcome}{numerical vector of the outcome}

\item{covariates}{data frame holding numerical covariates}

\item{modelCoef}{matrix of the model coefficients}

\item{fittedVals}{numerical vector of fitted values ('yhat')}

\item{residuals}{numerical vector of residuals}

\item{summaryStats}{list of summary statistics}

\item{leverage}{numerical vector of leverage of fitted values}
}
\description{
Initialiser for LinRegP objects
}

