% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\docType{class}
\name{LinRegP-class}
\alias{LinRegP-class}
\title{S4 class definition for a linear model}
\description{
S4 class definition for a linear model
}
\section{Slots}{

\describe{
\item{\code{model}}{character vector description of model}

\item{\code{intercept}}{is an intercept included in the model? Takes values TRUE or FALSE}

\item{\code{outcome}}{numeric vector of the outcome ('y')}

\item{\code{covariates}}{data frame of the covariates. At present must be numeric}

\item{\code{modelCoef}}{matrix representation of the fitted model coefficients ('betahats')}

\item{\code{fittedVals}}{numeric vector of the fitted values ('yhat')}

\item{\code{residuals}}{numeric vector of the fitted residuals}

\item{\code{summaryStats}}{list containing summary stats: Rsquared, weighted Rsquared and RMSE}

\item{\code{leverage}}{numeric vector of the leverage of each fitted value}
}}

