% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\name{lmp}
\alias{lmp}
\title{Constructor for LinRegP}
\usage{
lmp(outcome, covariates, int = TRUE, df)
}
\arguments{
\item{outcome}{character vector name of column containing outcome ("y")}

\item{covariates}{character vector of column names containing covariates ('x')}

\item{int}{logical. is an intercept included?}

\item{df}{data frame}
}
\description{
Constructor for LinRegP
}

