% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{plot,LinRegP-method}
\alias{plot,LinRegP-method}
\title{Plot generic. Note that do not use standardised residuals as the S3
lm class does}
\usage{
\S4method{plot}{LinRegP}(x, y, ...)
}
\arguments{
\item{x}{plotting param}

\item{y}{plotting param}

\item{...}{all other arguments}
}
\description{
Plot generic. Note that do not use standardised residuals as the S3
lm class does
}

