% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{predict,LinRegP-method}
\alias{predict,LinRegP-method}
\title{Predict generic}
\usage{
\S4method{predict}{LinRegP}(object, newdata, ...)
}
\arguments{
\item{object}{object of class LinRegP}

\item{newdata}{new data to predict outcome}

\item{...}{any other arguments}
}
\description{
Predict generic
}

