% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{show,LinRegP-method}
\alias{show,LinRegP-method}
\title{Show generic}
\usage{
\S4method{show}{LinRegP}(object)
}
\arguments{
\item{object}{object of class LinRegP}
}
\description{
Show generic
}

