% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{summary,LinRegP-method}
\alias{summary,LinRegP-method}
\title{Summary generic}
\usage{
\S4method{summary}{LinRegP}(object)
}
\arguments{
\item{object}{object of class LinRegP}
}
\description{
Summary generic
}

